nix: container build expression and SPDK changes
The script generates a ms-build container which will (should) use the
same nixpkgs as pinned with in this repo.

Note; that currently, some binary utilities are missing (i.e mkfs) this
is not done right now as we we still run CI on gitlab and we want to
first use some basic actions before we do the whole thing.

The added build.yaml runs the container generated by this script, runs
clippy and does a build. Note that clippy is run with nightly.

Additionally the SPDK changes will now also result in having some of
the test binaries available, like for example `spdk_tgt` which means
that overtime, we also remove the one we wrote ourselves. In particular
bdevperf which is getting verification capabilities will be useful.