test: fix mount_fs intermittent failures
As it turns out, threads spawned with std::thread::spawn() do not seem
to unaffinitize themselves. This brakes the assumption of having the
reactor make forward progress, after spawning a thread.

With this change, I have not seen any errors for this test, and I have
added a test-case that validates that threads spawned with
spawn_unaffinitized() are not running on a core we run our reactor(s)
on.

Note that, unlike the implementation of rte_xxx for setting thread
affinities, I'm using sched_setaffinity(). The reason for this is that
the former requires a pthread_t (which you dont get directly using
spawn()). The rte_xxx functions use pthread_set which itself is built on top of
sched_setaffinity()