test: add framework to start mayastor instances
Its been rather complicated to start/stop multiple mayastors within a
single test case. This is true for the BDD tests as well as the unit
tests.

The bookkeeping of the processes itself is a PITA, but things get more
complicated when we want to send signals, pause instances and
the likes. We need to be able to do this simply and easily for more
sophisticated recovery logic.

So using containers seems to be a logical, and instead of using YAML
files to compose targets, I've opted for using the docker remote API
directly. Note, this API  can be used on anything that runs docker
local or remote. However, we use "FROM SCRATCH" and do not make use of
a basic image, and start the artefact located in /target directly.

This docker API rather verbose, and so you will see alot of Some() and
Default but once you get over that is surprisingly well written. The
reason for not simply do a fork of compose up down is that we
need/want to "do things" with the containers as well.

Consider the case where we want to emulate a nexus loosing its
replica. To do this we would need to call pause -- but pause who?

What if we want to send a grpc command to one of the targets, send it
to who?

We could have gone the easy route, which is to hard code everything
but that would get us into duplicated constants between config files
and exec calls.

The test suites are now driven by tokio, and have a fundamental
difference compared to the previous test. Before this commit, when you
did something like mayastor_start(||  whatever()) the test thread
itself, was running mayastor. That now is changed, and the test thread
itself is not running mayastor and you keep control over it without
the need for polling or blocking calls at all.

An example is given in mayastor_compose_basic.rs and i've adapted an
existing test as well.