feat(poller): want poller abstraction
This change adds a safe abstraction around the poller interface. We
did not use pollers much (if not at all) but now that we need to
construct, among others our own qpairs we need to create them often
and frequently.

A small trampoline function (which is always inlined) is used to avoid
passing around raw pointers. The poller will execute a closure and allows
us to capture variables so there is no need for boxing arguments.

Implements CAS-610